+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.14.4] - 2021-02-09
+### Changes
+- Add app::event_mouse_button();
+- Add enums::Mouse.
+- Fix docs.
+- impl IntoIter for MenuItem.
 
 ## [0.14.3] - 2021-02-08
 ### Changes